Moved some tests to a different module and added an outline for more decorator tests to follow.
