+++ tests/test_enforce.py
@@ -28,33 +28,6 @@ class GeneralTests(unittest.TestCase):
 
         self.assertEqual(result, 0)
 
-    def test_class_decorator(self):
-        class SampleClass:
-            def test(self, data: int) -> int:
-                return data
-
-        sample = SampleClass()
-        self.assertEqual(sample.test(1), 1)
-
-        SampleClass = runtime_validation(SampleClass)
-        sample = SampleClass()
-        self.assertEqual(sample.test(1), 1)
-
-        with self.assertRaises(RuntimeTypeError):
-            sample.test('')
-
-    def test_method_decorator(self):
-        class SampleClass:
-            @runtime_validation
-            def test(self, data: int) -> int:
-                return data
-
-        sample = SampleClass()
-        self.assertEqual(sample.test(1), 1)
-
-        with self.assertRaises(RuntimeTypeError):
-            sample.test('')
-
     @runtime_validation
     def sample_function(self, text: str, data: typing.Union[int, None]) -> typing.Optional[int]:
         try:
